requirements.txt
Django==5.1.7
channels==4.2.2
channels-redis==4.3.0
daphne==4.2.0

psycopg2-binary==2.9.11
redis==7.3.0

python-dotenv==1.1.0
python-decouple==3.8

requests==2.32.3

django-cors-headers==4.7.0
django-axes==8.0.0
django-sequences==3.0

pillow==11.2.1

twilio==9.10.2
razorpay==1.4.2

cloudinary==1.44.1